requirements.txt
torch>=1.13.0
torchvision>=0.14.0
mmcv-full==1.7.0
mmdet==2.28.2
pycocotools
opencv-python
Pillow
clip-by-openai
segment-anything
tqdm
numpy